refactor(codegen): deduplicate inst results
