+++ crates/codegen/src/transform/load_pre.rs
@@ -79,6 +79,7 @@ use crate::{
         MirType, StorageAlias, Terminator, Value, ValueId,
     },
     pass::FunctionPass,
+    transform::inst_results,
     utils::repair_reachability_phis,
 };
 use alloy_primitives::U256;
@@ -515,7 +516,7 @@ impl LoadRedundancyEliminator {
             outs,
             reach,
             dominators: cfg.dominators().clone(),
-            inst_results: Self::inst_results(func),
+            inst_results: inst_results(func),
             inst_blocks: Self::inst_blocks(func),
         })
     }
@@ -1162,15 +1163,6 @@ impl LoadRedundancyEliminator {
             }
         }
         predecessors
-    }
-
-    fn inst_results(func: &Function) -> FxHashMap<InstId, ValueId> {
-        func.values
-            .iter_enumerated()
-            .filter_map(|(value_id, value)| {
-                if let Value::Inst(inst_id) = value { Some((*inst_id, value_id)) } else { None }
-            })
-            .collect()
     }
 
     fn inst_blocks(func: &Function) -> FxHashMap<InstId, BlockId> {
